- implemented the metadata handling for complex classes - refactored the runtime and split the base class and conditions from the utils module - renamed util functions to conform to `rune_...` naming if used by the generator
