+++ src/rune/runtime/utils.py
@@ -1,24 +1,15 @@
 '''Utility functions (runtime) for rune models.'''
 from __future__ import annotations
-import logging as log
 import keyword
 from enum import Enum
-from typing import get_args, get_origin
-from typing import TypeVar, Generic, Callable, Any
-from collections import defaultdict
-from pydantic import BaseModel, ValidationError, ConfigDict
+from typing import Callable, Any
 
 __all__ = [
-    'if_cond', 'if_cond_fn', 'Multiprop', 'rune_condition', 'BaseDataClass',
-    'ConditionViolationError', 'any_elements', 'get_only_element',
+    'if_cond', 'if_cond_fn', 'Multiprop', 'any_elements', 'get_only_element',
     'rune_filter', 'all_elements', 'contains', 'disjoint', 'join',
-    'rune_local_condition', 'execute_local_conditions', 'flatten_list',
-    'rune_resolve_attr', 'rune_count', 'rune_attr_exists', '_get_rune_object',
-    'set_rune_attr', 'add_rune_attr', 'check_cardinality', 'AttributeWithMeta',
-    'AttributeWithAddress', 'AttributeWithReference',
-    'AttributeWithMetaWithAddress', 'AttributeWithMetaWithReference',
-    'AttributeWithAddressWithReference',
-    'AttributeWithMetaWithAddressWithReference', 'rune_str'
+    'flatten_list', 'rune_resolve_attr', 'rune_count', 'rune_attr_exists',
+    '_get_rune_object', 'rune_set_attr', 'rune_add_attr', 'check_cardinality',
+    'rune_str'
 ]
 
 
@@ -40,14 +31,6 @@ def _to_list(obj) -> list | tuple:
     if isinstance(obj, (list, tuple)):
         return obj
     return (obj, )
-
-
-def _is_meta(obj: Any) -> bool:
-    '''Returns true if it is a meta data with embedded rune type.'''
-    return isinstance(
-        obj, (AttributeWithMeta, AttributeWithAddress,
-              AttributeWithMetaWithAddress, AttributeWithMetaWithReference,
-              AttributeWithMetaWithAddressWithReference))
 
 
 def mangle_name(attrib: str) -> str:
@@ -74,11 +57,6 @@ def rune_resolve_attr(obj: Any | None, attrib: str) -> Any | list[Any] | None:
             if item is not None
         ]
         return res if res else None
-    if _is_meta(obj):
-        # NOTE: ignores (for now) all meta attributes in the expressions.
-        # In the future one might want to check if the attrib is contained
-        # in the metadata and return it instead of failing.
-        obj = obj.value
     attrib = mangle_name(attrib)
     return getattr(obj, attrib, None)
 
@@ -128,283 +106,6 @@ class Multiprop(list):
             else:
                 res.append(getattr(x, attr))
         return res
-
-
-_CONDITIONS_REGISTRY: defaultdict[str, dict[str, Any]] = defaultdict(dict)
-
-
-def rune_condition(condition):
-    '''Wrapper to register all constraint functions in the global registry'''
-    path_components = condition.__qualname__.split('.')
-    path = '.'.join([condition.__module__ or ''] + path_components[:-1])
-    name = path_components[-1]
-    _CONDITIONS_REGISTRY[path][name] = condition
-
-    return condition
-
-
-def rune_local_condition(registry: dict):
-    '''Registers a condition function in a local registry.'''
-
-    def decorator(condition):
-        path_components = condition.__qualname__.split('.')
-        path = '.'.join([condition.__module__ or ''] + path_components)
-        registry[path] = condition
-
-        return condition
-
-    return decorator
-
-
-def execute_local_conditions(registry: dict, cond_type: str):
-    '''Executes all registered in a local registry.'''
-    for condition_path, condition_func in registry.items():
-        if not condition_func():
-            raise ConditionViolationError(
-                f"{cond_type} '{condition_path}' failed.")
-
-
-class ConditionViolationError(ValueError):
-    '''Exception thrown on violation of a constraint'''
-
-
-def _fqcn(cls) -> str:
-    return '.'.join([cls.__module__ or '', cls.__qualname__])
-
-
-def _get_conditions(cls) -> list:
-    res = []
-    index = cls.__mro__.index(BaseDataClass)
-    for c in reversed(cls.__mro__[:index]):
-        fqcn = _fqcn(c)
-        res += [('.'.join([fqcn, k]), v)
-                for k, v in _CONDITIONS_REGISTRY.get(fqcn, {}).items()]
-    return res
-
-
-class MetaAddress(BaseModel):  # pylint: disable=missing-class-docstring
-    scope: str | None = None
-    value: str
-
-
-class BaseDataClass(BaseModel):
-    ''' A base class for all cdm generated classes. It is derived from
-        `pydantic.BaseModel` which provides type checking at object creation
-        for all cdm classes. It provides as well the `validate_model`,
-        `validate_conditions` and `validate_attribs` methods which perform the
-        conditions, cardinality and type checks as specified in the rune
-        type model. The method `validate_model` is not invoked automatically,
-        but is left to the user to determine when to check the validity of the
-        cdm model.
-    '''
-    model_config = ConfigDict(extra='forbid',
-                              revalidate_instances='always',
-                              arbitrary_types_allowed=True)
-
-    meta: dict | None = None
-    address: MetaAddress | None = None
-
-    def validate_model(self,
-                       recursively: bool = True,
-                       raise_exc: bool = True,
-                       strict: bool = True) -> list:
-        ''' This method performs full model validation. It will validate all
-            attributes and it will also invoke `validate_conditions` to check
-            all conditions and the cardinality of all attributes of this object.
-            The parameter `raise_exc` controls whether an exception should be
-            thrown if a validation or condition is violated or if a list with
-            all encountered violations should be returned instead.
-        '''
-        att_errors = self.validate_attribs(raise_exc=raise_exc, strict=strict)
-        return att_errors + self.validate_conditions(recursively=recursively,
-                                                     raise_exc=raise_exc)
-
-    def validate_attribs(self,
-                         raise_exc: bool = True,
-                         strict: bool = True) -> list:
-        ''' This method performs attribute type validation.
-            The parameter `raise_exc` controls whether an exception should be
-            thrown if a validation or condition is violated or if a list with
-            all encountered violations should be returned instead.
-        '''
-        try:
-            self.model_validate(self, strict=strict)
-        except ValidationError as validation_error:
-            if raise_exc:
-                raise validation_error
-            return [validation_error]
-        return []
-
-    def validate_conditions(self,
-                            recursively: bool = True,
-                            raise_exc: bool = True) -> list:
-        ''' This method will check all conditions and the cardinality of all
-            attributes of this object. This includes conditions and cardinality
-            of properties specified in the base classes. If the parameter
-            `recursively` is set to `True`, it will invoke the validation on the
-            rune defined attributes of this object too.
-            The parameter `raise_exc` controls whether an exception should be
-            thrown if a condition is not met or if a list with all encountered
-            condition violations should be returned instead.
-        '''
-        self_rep = object.__repr__(self)
-        log.info('Checking conditions for %s ...', self_rep)
-        exceptions = []
-        for name, condition in _get_conditions(self.__class__):
-            log.info('Checking condition %s for %s...', name, self_rep)
-            if not condition(self):
-                msg = f'Condition "{name}" for {repr(self)} failed!'
-                log.error(msg)
-                exc = ConditionViolationError(msg)
-                if raise_exc:
-                    raise exc
-                exceptions.append(exc)
-            else:
-                log.info('Condition %s for %s satisfied.', name, self_rep)
-        if recursively:
-            for k, v in self.__dict__.items():
-                log.info('Validating conditions of property %s', k)
-                exceptions += _validate_conditions_recursively(
-                    v, raise_exc=raise_exc)
-        err = f'with {len(exceptions)}' if exceptions else 'without'
-        log.info('Done conditions checking for %s %s errors.', self_rep, err)
-        return exceptions
-
-    def check_one_of_constraint(self, *attr_names, necessity=True) -> bool:
-        """ Checks that one and only one attribute is set. """
-        values = self.model_dump()
-        vals = [values.get(n) for n in attr_names]
-        n_attr = sum(1 for v in vals if v is not None and v != [])
-        if necessity and n_attr != 1:
-            log.error('One and only one of %s should be set!', attr_names)
-            return False
-        if not necessity and n_attr > 1:
-            log.error('Only one of %s can be set!', attr_names)
-            return False
-        return True
-
-    def add_to_list_attribute(self, attr_name: str, value) -> None:
-        """
-        Adds a value to a list attribute, ensuring the value is of an allowed
-        type.
-
-        Parameters:
-        attr_name (str): Name of the list attribute.
-        value: Value to add to the list.
-
-        Raises:
-        AttributeError: If the attribute name is not found or not a list.
-        TypeError: If the value type is not one of the allowed types.
-        """
-        if not hasattr(self, attr_name):
-            raise AttributeError(f"Attribute {attr_name} not found.")
-
-        attr = getattr(self, attr_name)
-        if not isinstance(attr, list):
-            raise AttributeError(f"Attribute {attr_name} is not a list.")
-
-        # Get allowed types for the list elements
-        allowed_types = get_allowed_types_for_list_field(
-            self.__class__, attr_name)
-
-        # Check if value is an instance of one of the allowed types
-        if not isinstance(value, allowed_types):
-            raise TypeError(f"Value must be an instance of {allowed_types}, "
-                            f"not {type(value)}")
-
-        attr.append(value)
-
-
-def _validate_conditions_recursively(obj, raise_exc=True):
-    '''Helper to execute conditions recursively on a model.'''
-    if not obj:
-        return []
-    if isinstance(obj, BaseDataClass):
-        return obj.validate_conditions(
-            recursively=True,  # type:ignore
-            raise_exc=raise_exc)
-    if isinstance(obj, (list, tuple)):
-        exc = []
-        for item in obj:
-            exc += _validate_conditions_recursively(item, raise_exc=raise_exc)
-        return exc
-    if _is_meta(obj):
-        return _validate_conditions_recursively(obj.value, raise_exc=raise_exc)
-    return []
-
-
-def get_allowed_types_for_list_field(model_class: type, field_name: str):
-    """
-    Gets the allowed types for a list field in a Pydantic model, supporting
-    both Union and | operator.
-
-    Parameters:
-    model_class (type): The Pydantic model class.
-    field_name (str): The field name.
-
-    Returns:
-    tuple: A tuple of allowed types.
-    """
-    field_type = model_class.__annotations__.get(field_name)
-    if field_type and get_origin(field_type) is list:
-        list_elem_type = get_args(field_type)[0]
-        if get_origin(list_elem_type):
-            return get_args(list_elem_type)
-        return (list_elem_type, )  # Single type or | operator used
-    return ()
-
-
-ValueT = TypeVar('ValueT')
-
-
-class AttributeWithMeta(BaseModel, Generic[ValueT]):
-    '''Meta support'''
-    meta: dict | None = None
-    value: ValueT
-
-
-class AttributeWithAddress(BaseModel, Generic[ValueT]):
-    '''Meta support'''
-    address: MetaAddress | None = None
-    value: ValueT | None = None
-
-
-class AttributeWithReference(BaseDataClass):
-    '''Meta support'''
-    externalReference: str | None = None
-    globalReference: str | None = None
-
-
-class AttributeWithMetaWithAddress(BaseModel, Generic[ValueT]):
-    '''Meta support'''
-    meta: dict | None = None
-    address: MetaAddress | None = None
-    value: ValueT
-
-
-class AttributeWithMetaWithReference(BaseModel, Generic[ValueT]):
-    '''Meta support'''
-    meta: dict | None = None
-    externalReference: str | None = None
-    globalReference: str | None = None
-    value: ValueT
-
-
-class AttributeWithAddressWithReference(BaseModel, Generic[ValueT]):
-    '''Meta support'''
-    address: MetaAddress | None = None
-    externalReference: str | None = None
-    globalReference: str | None = None
-    value: ValueT
-
-
-class AttributeWithMetaWithAddressWithReference(BaseModel, Generic[ValueT]):
-    '''Meta support'''
-    meta: dict | None = None
-    address: MetaAddress | None = None
-    externalReference: str | None = None
-    globalReference: str | None = None
-    value: ValueT
 
 
 def _ntoz(v):
@@ -512,7 +213,7 @@ def rune_filter(items, filter_func, item_name='item'):
     return [item for item in items if filter_func(locals()[item_name])]
 
 
-def set_rune_attr(obj: Any, path: str, value: Any) -> None:
+def rune_set_attr(obj: Any, path: str, value: Any) -> None:
     """
     Sets an attribute of a rune model object to a specified value using a
     path.
@@ -550,7 +251,7 @@ def set_rune_attr(obj: Any, path: str, value: Any) -> None:
                              f"type {type(parent_obj).__name__}")
 
 
-def add_rune_attr(obj: Any, attrib: str, value: Any) -> None:
+def rune_add_attr(obj: Any, attrib: str, value: Any) -> None:
     """
     Adds a value to a list-like attribute of a rune model object.
